[ttl] Declare compute primitives through ODS
