+++ docs/sphinx/tour/index.md
@@ -118,7 +118,7 @@ A dataflow buffer is created with `ttl.make_dataflow_buffer_like` by passing a T
 
 The TT-NN tensor determines basic properties (likeness) such as data type and shape unit. The shape unit is a whole tile if the tensor has a tiled layout and is a scalar if the tensor has a row-major layout. Shape determines the shape of a block returned by one of the acquisition functions and is expressed in shape units. block count determines the total size of L1 memory allocated as a product of block size and block count. For the most common case block count defaults to 2 to enable double buffering.
 
-Physical tile dimensions supported by tt-metal have height 1, 2, 4, 8, 16, or 32 and width 16 or 32. Compute operations generally support 16x16, 16x32, 32x16, and 32x32 tiles. Elementwise operations and fills additionally support 1x32, 2x32, 4x32, and 8x32 tiles without matmul. Compute regions containing matmul retain support for these short-height tiles across the region. Matmul requires matching physical K dimensions and derives the result tile dimensions from the lhs height and rhs width. Current matmul LLKs accept 16x32, 32x16, or 32x32 rhs tiles and reject a 16x16 lhs. Transposed matmul also rejects a 32x16 rhs and the 32x32-lhs-by-16x32-rhs configuration. BFP compute operations currently require 32x32 tiles; data movement supports every physical tile dimension listed above.
+Physical tile dimensions supported by tt-metal have height 1, 2, 4, 8, 16, or 32 and width 16 or 32. Compute operations generally support 16x16, 16x32, 32x16, and 32x32 tiles. Elementwise operations, fills, and matmul additionally support 1x32, 2x32, 4x32, and 8x32 tiles. Each operation is validated independently, so fusion with a matmul does not extend another primitive's supported dimensions. Matmul requires matching physical K dimensions and derives the result tile dimensions from the lhs height and rhs width. Current matmul LLKs accept 16x32, 32x16, or 32x32 rhs tiles and reject a 16x16 lhs. Transposed matmul also rejects a 32x16 rhs and the 32x32-lhs-by-16x32-rhs configuration. BFP compute operations currently require 32x32 tiles; data movement supports every physical tile dimension listed above.
 
 ```{mermaid}
 graph LR
