+++ include/ttlang/Dialect/TTL/Transforms/ComputeTarget.h
@@ -6,6 +6,7 @@
 #define TTLANG_DIALECT_TTL_TRANSFORMS_COMPUTETARGET_H
 
 #include "ttlang/Dialect/TTCore/IR/TTCoreOpsTypes.h"
+#include "ttlang/Dialect/TTL/IR/TTL.h"
 
 #include "mlir/IR/Operation.h"
 #include "mlir/Support/LogicalResult.h"
@@ -14,23 +15,6 @@
 #include <string>
 
 namespace mlir::tt::ttl {
-
-/// Target-independent primitive used by compute capability queries.
-enum class ComputePrimitive {
-  Add,
-  Subtract,
-  Multiply,
-  ElementwiseBinary,
-  ElementwiseUnary,
-  Broadcast,
-  Reduce,
-  Transpose,
-  Fill,
-  Matmul,
-  Typecast,
-  MultiplyByConstant,
-  Passthrough,
-};
 
 /// Immutable LLK capabilities for one compute target.
 class ComputeTargetEnvironment {
@@ -51,7 +35,7 @@ public:
 
   virtual LogicalResult
   validatePrimitiveTileShape(ComputePrimitive primitive,
-                             ttcore::TileType tileType, bool containsMatmul,
+                             ttcore::TileType tileType,
                              std::string &failureReason) const = 0;
 
   virtual LogicalResult
@@ -63,7 +47,7 @@ public:
                           ttcore::TileType resultType, bool transposeRhs,
                           std::string &failureReason) const = 0;
 
-  LogicalResult validateOperation(Operation *operation, bool containsMatmul,
+  LogicalResult validateOperation(Operation *operation,
                                   std::string &failureReason) const;
 
 protected:
@@ -72,8 +56,6 @@ protected:
 
 /// Return the target-independent primitive implemented by a TTL operation.
 std::optional<ComputePrimitive> getComputePrimitive(Operation *operation);
-
-bool containsMatmulOperation(Operation *scope);
 
 } // namespace mlir::tt::ttl
